rename codegen files
