+++ package.json
@@ -235,10 +235,10 @@
         "RNSFormSheetContentWrapper": {
           "className": "RNSFormSheetContentWrapperComponentView"
         },
-        "RNSContainedModal": {
+        "RNSContainedModalHost": {
           "className": "RNSContainedModalHostComponentView"
         },
-        "RNSContainedModalProvider": {
+        "RNSContainedModalHostProvider": {
           "className": "RNSContainedModalProviderHostComponentView"
         }
       }
