+++ src/components/gamma/modals/contained-modal/ContainedModalProvider.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import ContainedModalProviderNativeComponent from '../../../../fabric/gamma/modals/contained-modal/ContainedModalProviderNativeComponent';
+import ContainedModalProviderHostNativeComponent from '../../../../fabric/gamma/modals/contained-modal/ContainedModalProviderHostNativeComponent';
 import type { ContainedModalProviderProps } from './ContainedModalProvider.types';
 
 export function ContainedModalProvider(props: ContainedModalProviderProps) {
-  return <ContainedModalProviderNativeComponent {...props} />;
+  return <ContainedModalProviderHostNativeComponent {...props} />;
 }
